fix(lints): warnings-as-errors clean across the workspace; restore eval fixtures
- Wire `[lints] workspace = true` into every crate so the workspace
  `warnings = "deny"` table actually applies; `clippy::all` is now denied
  in practice (CI runs clippy with `-D warnings`).
- Fix all rustc and clippy warnings workspace-wide (unused imports/vars,
  div_ceil, compound assign, needless borrows/casts, range-loop rewrites,
  derivable Default, doc-comment formatting, len-without-is_empty,
  checked_div, etc.). Behavior-preserving throughout.
- Wire the ROCm `gpu-test-shims` feature into the three GPU parity test
  targets via `required-features` and enable it in the CI GPU gate job,
  fixing E0433 compile errors.
- Real bugs found by the sweep: duplicated `acc = add_tensors(...)` line
  in grim-nn scythe2 device fan-in; missing `#[test]` on
  rms_norm_normalizes_to_unit_when_weight_is_one.
- Justified, narrowly-scoped `#[allow]`s only (FFI-shaped >7-arg kernel
  launchers, inherent trait-shadowing names kept for API compatibility,
  cubecl macro dialect limits), each with a one-line comment.
- grim-garage rocm: make `parse_rocminfo_text` a pure function (live
  sysfs telemetry moved to an enrich pass in `query_rocminfo_gpus`) —
  the fixture tests were environment-dependent and failed on real-GPU
  hosts. Relax the SoulEater dashboard assert (option gained a `title`
  attribute).
- Restore docs/eval fixtures (wikitext2 sample, golden baseline, gsm8k
  set) deleted in 2a46d77 while tests and the CI eval gate still
  reference them; fixes grim-cli recipe test and CI eval smoke.

Verification: `cargo check --workspace --all-targets` with
RUSTFLAGS="-D warnings" → 0; `cargo clippy --workspace --all-targets
--exclude vulkan --exclude metal -- -D warnings` → 0; `cargo fmt --all
-- --check` → clean; 1559 workspace lib tests pass; backend parity CPU
leg and doctests pass.

+++ crates/grim-backend-rocm/tests/attn_o_proj_fusion.rs
@@ -17,7 +17,7 @@ fn gpu_device() -> Option<RocmDevice> {
 
 fn fill(n: usize, seed: f32) -> Vec<f32> {
     (0..n)
-        .map(|i| ((seed + i as f32 * 0.61).sin() * 0.5 - 0.1) as f32)
+        .map(|i| (seed + i as f32 * 0.61).sin() * 0.5 - 0.1)
         .collect()
 }
